Tenatatively bump to 0.8
This branch is based off of the ghc8 branch upstream. Until that gets
released, we'll be keeping this fork around, but I wanted to bump a
breaking version number since APIs changed. It is likely that when
upstream changes, it will take on the 0.8 set of versions anyways, so
we won't have to go around bumping versions around when that happens.